+++ src/python_takserver_api/tak_class.py
@@ -25,6 +25,7 @@ from .tak_file_user_account_management_api import UserAccountManagementApi
 from .tak_group_api import GroupApi
 from .tak_data_feed_api import DataFeedApi
 from .tak_cert_manager_api import CertManagerApi
+from .tak_submission_api import SubmissionApi
 from .tak_mission_api import MissionApi
 
 
@@ -51,6 +52,7 @@ class Server:
         self.groups = GroupApi(self)
         self.datafeeds = DataFeedApi(self)
         self.certs = CertManagerApi(self)
+        self.submission = SubmissionApi(self)
         self.mission = MissionApi(self)
 
     async def close(self) -> None: